Cluster Working with Functions
still using bandaid fix. Complete overhaul of planar segregation is up next.
